fix shit

+++ src/modules/logs.js
@@ -2,7 +2,6 @@ const Eris = require("eris");
 const threadUtils = require("../threadUtils");
 const threads = require("../data/threads");
 const moment = require("moment");
-const config = require("../config");
 const utils = require("../utils");
 
 /**
@@ -14,7 +13,14 @@ module.exports = bot => {
      * @param {String} userId
      */
     async function getLogs(userId) {
-      const userThreads = await threads.getClosedThreadsByUserId(userId);
+      let userThreads = await threads.getClosedThreadsByUserId(userId);
+
+      if (! utils.isAdmin(msg.member)) {
+        userThreads = userThreads.filter((t) => ! t.isPrivate);
+      }
+
+      if (! userThreads.length) return bot.createMessage(msg.channel.id, "No logs found.");
+
       // Descending by date
       userThreads.sort((a, b) => {
         if (a.created_at > b.created_at) return -1;
@@ -24,11 +30,9 @@ module.exports = bot => {
 
       const threadLines = await Promise.all(userThreads.map(async thread => {
         const logUrl = await thread.getLogUrl();
-        const formattedDate = moment.utc(thread.created_at).format("MMM Do YYYY [at] HH:mm [UTC]");
-        return `\`${formattedDate}\`: <${logUrl}>`;
+        const formattedDate = moment.utc(thread.created_at).format("YYYY-MM-DD HH:mm [UTC]");
+        return `\`${formattedDate}\` ${thread.scheduled_close_name}: <${logUrl}>`;
       }));
-
-      if(! userThreads || ! userThreads.length) return msg.channel.createMessage("No logs found.");
       const message = `**Log files for <@${userId}>:**\n${threadLines.join("\n")}`;
 
       // Send the list of logs in chunks of 15 lines per message
@@ -40,7 +44,7 @@ module.exports = bot => {
        */
       let root = Promise.resolve();
       chunks.forEach(lines => {
-        root = root.then(() => msg.channel.createMessage(lines.join("\n")));
+        root = root.then(() => bot.createMessage(msg.channel.id, lines.join("\n")));
       });
     }
 
@@ -49,31 +53,27 @@ module.exports = bot => {
      */
     async function deleteLogs(userId) {
       await threads.deleteClosedThreadsByUserId(userId);
-      msg.channel.createMessage(`Deleted log files for <@!${userId}>`);
+      bot.createMessage(msg.channel.id, `Deleted log files for <@!${userId}>`);
     }
+
+    let userId = thread && thread.user_id;
 
     if (args.length > 0) {
-      if (args[0] === "delete") {
+      if (args[0] === "delete" && utils.isAdmin(msg.member)) {
         const userId = utils.getUserMention(args.slice(1).join(" "));
-        if (! userId) return;
+        if (! userId) return utils.postSystemMessageWithFallback(msg.channel, thread, "Please provide a user mention or ID!");
 
-        if (! config.inboxAdminRoleId) {
-          return;
-        }
-
-        if (msg.member.roles && msg.member.roles.includes(config.inboxAdminRoleId)) {
-          deleteLogs(userId);
-        }
+        return deleteLogs(userId);
       }
 
-      // User mention/id as argument
-      const userId = utils.getUserMention(args.join(" "));
-      if (! userId) return;
-      getLogs(userId);
-    } else if (thread) {
-      // Calling !logs without args in a modmail thread returns the logs of the user of that thread
-      getLogs(thread.user_id);
+      userId = utils.getUserMention(args.join(" "));
     }
+
+    if (! userId) return utils.postSystemMessageWithFallback(msg.channel, thread, "Please provide a user mention or ID!");
+
+    // Calling !logs without args in a modmail thread returns the logs of the user of that thread
+
+    return getLogs(userId);
   });
 
   threadUtils.addInboxServerCommand(bot, "loglink", async (msg, args, thread) => {
